Changed to struct variants
Closes https://github.com/PistonDevelopers/input/issues/27
